server/requirements.txt
paho-mqtt
python-dotenv
keyboard
